feat: Add identity service and connect to todolist service

+++ TodoList.Service/Infrastructure/Repositories/TodoRepository.cs
@@ -1,6 +1,7 @@
 using Microsoft.EntityFrameworkCore;
-using TodoList.Service.Domain;
+using TodoList.Service.Domain.Entities;
 using TodoList.Service.Domain.Repositories;
+using TodoList.Service.Infrastructure.Contexts;
 
 namespace TodoList.Service.Infrastructure.Repositories;
 
@@ -13,27 +14,35 @@ public class TodoRepository : ITodoRepository
     this.db = db;
   }
 
-  public async Task<Todo> GetAsync(int id)
+  public async Task<Todo> GetAsync(int id, string userId)
   {
-    return await db.Todos.FindAsync(id) is Todo todo ? todo : Todo.NullTodo;
+    if (await db.Todos.FindAsync(id) is Todo todo)
+    {
+      return todo.UserId == userId ? todo : Todo.NullTodo;
+    }
+
+    return Todo.NullTodo;
   }
 
-  public async Task<IEnumerable<Todo>> GetAllAsync()
+  public async Task<IEnumerable<Todo>> GetAllAsync(string userId)
   {
-    return await db.Todos.ToListAsync();
+    return await db.Todos.Where(t => t.UserId == userId).ToListAsync();
   }
 
-  public async Task<Todo> CreateAsync(Todo todo)
+  public async Task<Todo> CreateAsync(Todo todo, string userId)
   {
+    todo.UserId = userId;
+
     db.Todos.Add(todo);
     await db.SaveChangesAsync();
+
     return todo;
   }
 
-  public async Task<Todo> UpdateAsync(int id, Todo newTodo)
+  public async Task<Todo> UpdateAsync(int id, Todo newTodo, string userId)
   {
     var todo = await db.Todos.FindAsync(id);
-    if (todo is null)
+    if (todo is null || todo.UserId != userId)
     {
       return Todo.NullTodo;
     }
@@ -45,9 +54,9 @@ public class TodoRepository : ITodoRepository
     return todo;
   }
 
-  public async Task<Todo> DeleteAsync(int id)
+  public async Task<Todo> DeleteAsync(int id, string userId)
   {
-    if (await db.Todos.FindAsync(id) is Todo todo)
+    if (await db.Todos.FindAsync(id) is Todo todo && todo.UserId == userId)
     {
       db.Todos.Remove(todo);
       await db.SaveChangesAsync();
